pyproject.toml
[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[project]
name = "builder"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
  "actions-python-core",
  "actions-python-github[githubkit]",
  "packaging",
  "ruamel.yaml",
]

[project.scripts]
build-container = "builder.container:build_container"
build-go = "builder.go:build_go"
build-package = "builder.package:build_package"
matrix = "builder.matrix:matrix"

[tool.ruff.lint]
select = ["F", "E", "W", "I", "UP"]
ignore = ["E501"]

[tool.uv.build-backend]
module-root = ""
module-name = "builder"
